chore(engine): simplify cache and persistence wait logic in reth_newPayload
- Replace broken spawn_blocking + try_recv parallel wait with sequential
  blocking: wait for persistence first, then caches. The previous
  implementation would silently lose persistence results if they took
  longer than cache waits.
- Simplify PayloadProcessor::wait_for_caches from spawn_blocking +
  channels to direct sequential calls.
- Move CacheWaitDurations and WaitForCaches trait from
  payload_processor to tree module (bottom of file).
- Make wait_for_caches a regular method on PayloadProcessor instead of
  a trait impl, since only BasicEngineValidator needs the trait bound.
- Remove unnecessary temporary variable for payload_validator reference.
- Fix 'Peresistence' typo.

Amp-Thread-ID: https://ampcode.com/threads/T-019c662e-9052-716a-a3c0-1f1c2d7d4dc4
